[2, 4단계 방탈출 결제/배포] 명오(김형호) 미션 제출합니다. (#81)
* style: 테스트 함수명 변경

* test: mock이 아닌 fake 사용

* test: PaymentApprove 테스트 추가

* refactor: 단순 널 체킹에 사용되는 optional 제거

* refactor: 예외 메시지가 외부 API에 의존하지 않도록 변경 및 log 추가

* refactor: 테이블 구조 변경 및 웹에서 paymentKey와 amount를 응답하도록 수정

* refactor: restDoc을 이용한 api 명세서 추가

* chore: gitignore가 adoc을 무시하도록 수정

* refactor: 로깅 삭제

* refactor: swagger 제거

* refactor: 타임아웃 시간 수정

* feat: asciidocs 문서 추가

* style: 생성자 파라미터 개행 추가

* refactor: 결제 상태 중 계좌이체 제거

* refactor: 불필요한 생성자 제거

* fix: db 데이터 오류 수정

* refactor: 중복 로직 제거 및 api 문서 소폭 수정("*" 제거)

* refactor: 베포 환경에서 mysql을 사용하도록 수정

* refactor: 로컬에서는 h2 데이터베이스를 사용하도록 수정

* refactor: production 환경과 develop 환경 property 분리 및 민감한 상수 별도로 관리

* docs: snippet의 식별자 수정 및 문서 이름 변경

* feat: 외부 api 및 예기치 못한 오류에 대한 log 추가

* docs: gitignore 수정 및 data.sql 제거

---------

+++ src/main/java/roomescape/core/dto/reservation/ReservationPaymentRequest.java
@@ -22,14 +22,14 @@ public class ReservationPaymentRequest {
     private String orderId;
 
     @NotNull(message = "결제 금액은 null일 수 없습니다.")
-    private Integer amount;
+    private Long amount;
 
     public ReservationPaymentRequest() {
     }
 
     public ReservationPaymentRequest(final String date, final Long timeId, final Long themeId,
                                      final String paymentKey, final String orderId,
-                                     final Integer amount) {
+                                     final Long amount) {
         this.date = date;
         this.timeId = timeId;
         this.themeId = themeId;
@@ -58,7 +58,7 @@ public class ReservationPaymentRequest {
         return orderId;
     }
 
-    public Integer getAmount() {
+    public Long getAmount() {
         return amount;
     }
 }
